requirements.txt
# Windows 10 Antivirus - Requirements
# Install: pip install -r requirements.txt

# GUI Framework
customtkinter>=5.2.0
pystray>=0.19.5
Pillow>=10.0.0

# File System Monitoring
watchdog>=3.0.0

# Network & System
psutil>=5.9.0
scapy>=2.5.0
pywin32>=306
wmi>=1.5.1

# Security & Cryptography
cryptography>=41.0.0
pycryptodome>=3.19.0

# Database
# SQLite is built into Python

# Scheduling
schedule>=1.2.0
APScheduler>=3.10.0

# Logging
loguru>=0.7.0

# HTTP Requests (for updates, phishing checks)
requests>=2.31.0
aiohttp>=3.9.0

# Email Processing
imapclient>=2.3.1
email-validator>=2.1.0

# Compression for backups
py7zr>=0.20.0
